+++ src/type/integer_type.cpp
@@ -48,7 +48,7 @@ Value IntegerType::Add(const Value& left, const Value &right) const {
   case Type::BIGINT:
     return AddValue<int32_t, int64_t>(left, right);
   case Type::DECIMAL:
-    return ValueFactory::GetDoubleValue(left.value_.integer + right.GetAs<double>());
+    return ValueFactory::GetDecimalValue(left.value_.integer + right.GetAs<double>());
   default:
     break;
   }
@@ -73,7 +73,7 @@ Value IntegerType::Subtract(const Value& left, const Value &right) const {
   case Type::BIGINT:
     return SubtractValue<int32_t, int64_t>(left, right);
   case Type::DECIMAL:
-    return ValueFactory::GetDoubleValue(left.value_.integer - right.GetAs<double>());
+    return ValueFactory::GetDecimalValue(left.value_.integer - right.GetAs<double>());
   default:
     break;
   }
@@ -98,7 +98,7 @@ Value IntegerType::Multiply(const Value& left, const Value &right) const {
   case Type::BIGINT:
     return MultiplyValue<int32_t, int64_t>(left, right);
   case Type::DECIMAL:
-    return ValueFactory::GetDoubleValue(left.value_.integer * right.GetAs<double>());
+    return ValueFactory::GetDecimalValue(left.value_.integer * right.GetAs<double>());
   default:
     break;
   }
@@ -127,7 +127,7 @@ Value IntegerType::Divide(const Value& left, const Value &right) const {
   case Type::BIGINT:
     return DivideValue<int32_t, int64_t>(left, right);
   case Type::DECIMAL:
-    return ValueFactory::GetDoubleValue(left.value_.integer / right.GetAs<double>());
+    return ValueFactory::GetDecimalValue(left.value_.integer / right.GetAs<double>());
   default:
     break;
   }
@@ -156,7 +156,7 @@ Value IntegerType::Modulo(const Value& left, const Value &right) const {
   case Type::BIGINT:
     return ModuloValue<int32_t, int64_t>(left, right);
   case Type::DECIMAL:
-    return ValueFactory::GetDoubleValue(
+    return ValueFactory::GetDecimalValue(
         ValMod(left.value_.integer, right.GetAs<double>()));
   default:
     break;
@@ -168,13 +168,13 @@ Value IntegerType::Modulo(const Value& left, const Value &right) const {
 Value IntegerType::Sqrt(const Value& val) const {
   PL_ASSERT(val.CheckInteger());
   if (val.IsNull())
-    return ValueFactory::GetDoubleValue(PELOTON_DECIMAL_NULL);
+    return ValueFactory::GetDecimalValue(PELOTON_DECIMAL_NULL);
 
   if (val.value_.integer < 0) {
     throw Exception(EXCEPTION_TYPE_DECIMAL,
         "Cannot take square root of a negative number.");
   }
-  return ValueFactory::GetDoubleValue(sqrt(val.value_.integer));
+  return ValueFactory::GetDecimalValue(sqrt(val.value_.integer));
 
 }
 
@@ -189,7 +189,7 @@ Value IntegerType::OperateNull(const Value& left UNUSED_ATTRIBUTE, const Value &
   case Type::BIGINT:
     return ValueFactory::GetBigIntValue((int64_t) PELOTON_INT64_NULL);
   case Type::DECIMAL:
-    return ValueFactory::GetDoubleValue((double) PELOTON_DECIMAL_NULL);
+    return ValueFactory::GetDecimalValue((double) PELOTON_DECIMAL_NULL);
   default:
     break;
   }
@@ -465,8 +465,8 @@ Value IntegerType::CastAs(const Value& val, const Type::TypeId type_id) const {
   }
   case Type::DECIMAL: {
     if (val.IsNull())
-      return ValueFactory::GetDoubleValue(PELOTON_DECIMAL_NULL);
-    return ValueFactory::GetDoubleValue((double) val.GetAs<int32_t>());
+      return ValueFactory::GetDecimalValue(PELOTON_DECIMAL_NULL);
+    return ValueFactory::GetDecimalValue((double) val.GetAs<int32_t>());
   }
   case Type::VARCHAR:
     if (val.IsNull())
