+++ test/common/container_tuple_test.cpp
@@ -33,7 +33,7 @@ TEST_F(ContainerTupleTest, VectorValue) {
   std::vector<type::Value> values;
   values.push_back(type::ValueFactory::GetIntegerValue(11));
   values.push_back(type::ValueFactory::GetIntegerValue(22));
-  values.push_back(type::ValueFactory::GetDoubleValue(3.14));
+  values.push_back(type::ValueFactory::GetDecimalValue(3.14));
   values.push_back(type::ValueFactory::GetVarcharValue("Hello from ContainerTupleTest"));
 
   expression::ContainerTuple<std::vector<type::Value>> ctuple(&values);
